1. resolve 1600 compile warns.

+++ ClouSoft/app/DataBase/OIObjInfo.h
@@ -212,7 +212,7 @@ int InputCtrlMethod6(WORD wOI, BYTE bMethod, BYTE bOpMode, BYTE* pbPara, int iPa
 int QuitCtrlMethod7(WORD wOI, BYTE bMethod, BYTE bOpMode, BYTE* pbPara, int iParaLen, void* pvAddon, BYTE* pFmt, WORD wFmtLen, BYTE* pbRes, int* piRetLen=NULL);
 int InputCtrlMethod127(WORD wOI, BYTE bMethod, BYTE bOpMode, BYTE* pbPara, int iParaLen, void* pvAddon, BYTE* pFmt, WORD wFmtLen, BYTE* pbRes, int* piRetLen=NULL);
 
-BYTE GetOiClass(WORD wOI);
+WORD GetOiClass(WORD wOI);
 
 BYTE* GetSchFmt(BYTE bSchType, WORD* pwFmtLen);
